#51-nxt964-build-search-page-ui
This work not only Builds Search Page UI but also includes some changes at these file (to link pages):
- modify index.js of JobItem
- modify index.js of JobDetail
- modify index.js of ProfileDashboard
- modify JobItemSearch.module.css
- modify JobDetail.module.css
- modify ProfileDashboard.module.css
