+++ talenthive-website/frontend/src/components/JobItem/index.js
@@ -1,4 +1,4 @@
-import { useNavigate } from 'react-router';
+import { useNavigate, useNavigation } from 'react-router';
 import styleSearch from '../../styles/components/JobItemSearch.module.css'
 import styleHome from '../../styles/components/JobItemHome.module.css'
 import styleDetail from '../../styles/components/JobItemDetail.module.css'
@@ -44,10 +44,9 @@ const JobItem = {
     },
 
     ////// SEARCH PAGE //////
-    SearchPage: ({props, state}) => {
-        const navigate = useNavigate()
+    SearchPage: ({props, state, selected}) => {
         return ( 
-            <div className={styleSearch.wrapper} onClick={() => navigate('/job-detail')}>
+            <div className={`${styleSearch.wrapper} ${selected===true && styleSearch.selected}`}>
                 <div className={styleSearch.header}>
                     <p className={styleSearch.position}>{props.position}</p>
                     {state && <p className={styleSearch[state]}>{state}</p>}
@@ -80,14 +79,19 @@ const JobItem = {
 
     ////// HIRE TALENT PAGE //////
     Detail: ({props, isEmployer}) => {
-        const navigate = useNavigate()
+        const navigate = useNavigate();
+
+        const handleNavigate = () => {
+            navigate('/job-detail', {state: props});
+        }
+
         return ( 
-            <div className={styleDetail.wrapper} onClick={() => navigate('/job-detail')}>
+            <div className={styleDetail.wrapper} onClick={handleNavigate}>
                 <div className={styleDetail.header}>
                     <p className={styleDetail.position}>{props.position}</p>
                     {isEmployer && <p className={styleDetail[props.state]}>{props.state}</p>}
                 </div>
-                
+            
                 <div className={styleDetail.description}>
                     <div className={styleDetail.salary}>
                         <FontAwesomeIcon icon={faCircleDollarToSlot} className={styleDetail.icon}/>
