+++ talenthive-website/frontend/src/styles/components/JobItemSearch.module.css
@@ -15,6 +15,11 @@
 }
 
 .wrapper:hover {
+    cursor: pointer;
+    background-color: #f3f5f7;   
+}
+
+.selected {
     background-color: #f3f5f7;
     cursor: pointer;
     border: 2px solid var(--primary-color);
