+++ talenthive-website/frontend/src/styles/pages/ProfileDashboard.module.css
@@ -28,8 +28,8 @@
 .avatar {
     height: 100%;
     width: calc(var(--height-cover-page) - 2*var(--padding-cover-page));;
-    background-color: #ccc;
     border-radius: 4px;
+    object-fit: contain;
 }
 
 .name {
